Rename the cart.js to cartUtils.js
